MERN STACK WEB APPLICATION
The Job Posting Board is a full-stack web application that allows companies to register, post job openings, and manage job applications. Job seekers can view available job listings, apply for jobs, and receive alerts for new postings. This application is designed to streamline the job application process for both employers and job seekers, making it easier to find the right candidates and opportunities.